% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts_S3.R
\name{coef.shorts_mixed_model}
\alias{coef.shorts_mixed_model}
\title{S3 method for extracting model parameters from \code{shorts_mixed_model} object}
\usage{
\method{coef}{shorts_mixed_model}(object, ...)
}
\arguments{
\item{object}{\code{shorts_mixed_model} object}

\item{...}{Extra arguments. Not used}
}
\description{
S3 method for extracting model parameters from \code{shorts_mixed_model} object
}
\examples{
data("split_times")

mixed_model <- mixed_model_using_splits(
  data = split_times,
  distance = "distance",
  time = "time",
  athlete = "athlete"
)

# mixed_model$parameters
coef(mixed_model)
}
