% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-distance-time.R
\name{model_distance_time}
\alias{model_distance_time}
\title{@rdname model_functions
 @description \code{model_distance_time} estimates short sprint parameters using distance-time trace
     (e.g., timing gates/photo cells)}
\usage{
model_distance_time(distance, time, weights = 1, CV = NULL, na.rm = FALSE, ...)
}
\description{
@rdname model_functions
 @description \code{model_distance_time} estimates short sprint parameters using distance-time trace
     (e.g., timing gates/photo cells)
}
\examples{
split_distances <- c(10, 20, 30, 40, 50)
split_times <- create_timing_gates_splits(
  gates = split_distances,
  MSS = 10,
  MAC = 9,
  FD = 0,
  TC = 0
)

# Simple model
simple_model <- model_distance_time(split_distances, split_times)

print(simple_model)
coef(simple_model)
plot(simple_model)
 @export
}
