% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-distance-time.R
\name{model_distance_time_FD}
\alias{model_distance_time_FD}
\title{@rdname model_functions
 @description \code{model_distance_time_FD} estimates short sprint parameters using distance-time trace
     (e.g., timing gates/photo cells), with additional flying distance correction parameter \code{FD}}
\usage{
model_distance_time_FD(
  distance,
  time,
  weights = 1,
  FD = NULL,
  CV = NULL,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{FD}{Use this parameter if you do not want the \code{FD} parameter to be estimated, but rather take the
provided value}
}
\description{
@rdname model_functions
 @description \code{model_distance_time_FD} estimates short sprint parameters using distance-time trace
     (e.g., timing gates/photo cells), with additional flying distance correction parameter \code{FD}
}
\examples{
split_distances <- c(10, 20, 30, 40, 50)
split_times <- create_timing_gates_splits(
  gates = split_distances,
  MSS = 10,
  MAC = 9,
  FD = 0.5,
  TC = 0,
  noise = 0.001
)

# FD model
FD_model <- model_distance_time_FD(split_distances, split_times)

print(FD_model)
coef(FD_model)
plot(FD_model)
 @export
}
