% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-distance-time.R
\name{model_distance_time_TC}
\alias{model_distance_time_TC}
\title{@rdname model_functions
 @description \code{model_distance_time_TC} estimates short sprint parameters using distance-time trace
     (e.g., timing gates/photo cells), with additional time correction parameter \code{TC}}
\usage{
model_distance_time_TC(
  distance,
  time,
  weights = 1,
  CV = NULL,
  na.rm = FALSE,
  ...
)
}
\description{
@rdname model_functions
 @description \code{model_distance_time_TC} estimates short sprint parameters using distance-time trace
     (e.g., timing gates/photo cells), with additional time correction parameter \code{TC}
}
\examples{
split_distances <- c(10, 20, 30, 40, 50)
split_times <- create_timing_gates_splits(
  gates = split_distances,
  MSS = 10,
  MAC = 9,
  FD = 0,
  TC = 0.2,
  noise = 0.001
)

# TC model
TC_model <- model_distance_time_TC(split_distances, split_times)

print(TC_model)
coef(TC_model)
plot(TC_model)
 @export
}
