% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_using_splits.R
\name{model_split_times}
\alias{model_split_times}
\alias{model_using_splits}
\alias{model_using_splits_with_time_correction}
\alias{model_using_splits_with_corrections}
\title{Models Using Split Times}
\usage{
model_using_splits(
  distance,
  time,
  time_correction = 0,
  weights = 1,
  LOOCV = FALSE,
  na.rm = FALSE,
  ...
)

model_using_splits_with_time_correction(
  distance,
  time,
  weights = 1,
  LOOCV = FALSE,
  na.rm = FALSE,
  ...
)

model_using_splits_with_corrections(
  distance,
  time,
  weights = 1,
  LOOCV = FALSE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{distance, time}{Numeric vector. Indicates the position of the timing gates and time measured}

\item{time_correction}{Numeric vector. Used to correct for different starting techniques. This correction is
done by adding \code{time_correction} to \code{time}. Default is 0. See more in Haugen et al. (2018)}

\item{weights}{Numeric vector. Default is vector of 1.
This is used to give more weight to particular observations. For example, use \code{1\\distance} to give
more weight to observations from shorter distances.}

\item{LOOCV}{Should Leave-one-out cross-validation be used to estimate model fit? Default is \code{FALSE}}

\item{na.rm}{Logical. Default is FALSE}

\item{...}{Forwarded to \code{\link[stats]{nls}} function}
}
\value{
List object with the following elements:
    \describe{
        \item{parameters}{List with the following estimated parameters:
            \code{MSS}, \code{TAU}, \code{MAC}, \code{PMAX}, \code{time_correction}, and
            \code{distance_correction}}
        \item{model_fit}{List with the following components:
            \code{RSE}, \code{R_squared}, \code{minErr}, \code{maxErr}, and \code{RMSE}}
        \item{model}{Model returned by the \code{\link[stats]{nls}} function}
        \item{data}{Data frame used to estimate the sprint parameters, consisting of \code{distance},
            \code{time}, \code{weights}, and \code{pred_time} columns}
        }
}
\description{
These functions model the sprint split times using mono-exponential equation, where \code{time}
 is used as target or outcome variable, and \code{distance} as predictor. Function
 \code{\link{model_using_splits}} provides the simplest model with estimated \code{MSS} and \code{TAU}
 parameters. Time correction using heuristic rule of thumbs (e.g., adding 0.3s to split times) can be
 implemented using \code{time_correction} function parameter. Function
 \code{\link{model_using_splits_with_time_correction}}, besides estimating \code{MSS} and \code{TAU},
 estimates additional parameter \code{time_correction}.  Function \code{\link{model_using_splits_with_corrections}},
 besides estimating \code{MSS}, \code{TAU} and \code{time_correction}, estimates additional parameter
 \code{distance_correction}. For more information about these function please refer to accompanying vignettes in
 this package.
}
\examples{
split_times <- data.frame(
  distance = c(5, 10, 20, 30, 35),
  time = c(1.20, 1.96, 3.36, 4.71, 5.35)
)

# Simple model
simple_model <- with(
  split_times,
  model_using_splits(distance, time)
)

print(simple_model)
coef(simple_model)
plot(simple_model)

# Model with correction of 0.3s
model_with_correction <- with(
  split_times,
  model_using_splits(distance, time, time_correction = 0.3)
)

print(model_with_correction)
coef(model_with_correction)
plot(model_with_correction)

# Model with time_correction estimation
model_with_time_correction_estimation <- with(
  split_times,
  model_using_splits_with_time_correction(distance, time)
)

print(model_with_time_correction_estimation)
coef(model_with_time_correction_estimation)
plot(model_with_time_correction_estimation)

# Model with time and distance correction estimation
model_with_time_distance_correction_estimation <- with(
  split_times,
  model_using_splits_with_corrections(distance, time)
)

print(model_with_time_distance_correction_estimation)
coef(model_with_time_distance_correction_estimation)
plot(model_with_time_distance_correction_estimation)
}
\references{
Haugen TA, Tønnessen E, Seiler SK. 2012. The Difference Is in the Start: Impact of Timing and Start
        Procedure on Sprint Running Performance: Journal of Strength and Conditioning Research 26:473–479.
        DOI: 10.1519/JSC.0b013e318226030b.
}
