% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_using_split_times.R
\name{model_using_split_times_with_corrections}
\alias{model_using_split_times_with_corrections}
\title{Model Using Split Times with Corrections}
\usage{
model_using_split_times_with_corrections(
  distance,
  time,
  weights = 1,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{distance}{Numeric vector}

\item{time}{Numeric vector}

\item{weights}{Numeric vector. Default is vector of 1
This is used to give more weight to particular observations. For example, use \code{1\\distance} to give
more weight to observations from shorter distances.}

\item{na.rm}{Logical. Default is FALSE}

\item{...}{Forwarded to \code{\link[stats]{nls}} function}
}
\value{
List object with the following elements:
    \describe{
        \item{parameters}{List with the following estimated parameters:
            \code{MSS}, \code{TAU}, \code{MAC}, \code{PMAX}, \code{time_correction}, and
            \code{distance_correction}}
        \item{model_fit}{List with the following components:
            \code{RSE}, \code{R_squared}, \code{minErr}, \code{maxErr}, and \code{RMSE}}
        \item{model}{Model returned by the \code{\link[stats]{nls}} function}
        \item{data}{Data frame used to estimate the sprint parameters, consisting of \code{distance},
            \code{time}, \code{weights},and \code{pred_time} columns}
        }
}
\description{
This function models the sprint split times using mono-exponential equation that estimates
    maximum sprinting speed (\code{MSS}), relative acceleration (\code{TAU}), \code{time_correction},
    and \code{distance_correcion}. \code{time} is used as target or outcome variable, and \code{distance}
    as predictor.
}
\examples{
split_times <- data.frame(
distance = c(5, 10, 20, 30, 35),
time = c(1.20, 1.96, 3.36, 4.71, 5.35)
)

sprint_model <- with(
  split_times,
  model_using_split_times_with_corrections(distance, time)
)

sprint_model$parameters
}
