% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts_S3.R
\name{plot.shorts_mixed_model}
\alias{plot.shorts_mixed_model}
\title{S3 method for plotting \code{shorts_mixed_model} object}
\usage{
\method{plot}{shorts_mixed_model}(x, type = NULL, ...)
}
\arguments{
\item{x}{\code{shorts_mixed_model} object}

\item{type}{Not used}

\item{...}{Not used}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
S3 method for plotting \code{shorts_mixed_model} object
}
\examples{
# Split times
data("split_times")
mixed_model_splits <- mixed_model_using_splits(
  data = split_times,
  distance = "distance",
  time = "time",
  athlete = "athlete"
)

plot(mixed_model_splits)

# Radar gun data
data("radar_gun_data")
mixed_model_radar <- mixed_model_using_radar(radar_gun_data, "time", "velocity", "athlete")

plot(mixed_model_radar)
}
