% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts_S3.R
\name{predict.shorts_model}
\alias{predict.shorts_model}
\title{S3 method for predicting kinematics from \code{shorts_model} object}
\usage{
\method{predict}{shorts_model}(object, ...)
}
\arguments{
\item{object}{\code{shorts_model} object}

\item{...}{Extra arguments. Not used}
}
\value{
Data frame
}
\description{
Predicts sprint kinematics for 0-6sec (100hz) which include distance,
    velocity, acceleration, and relative power
}
\examples{
split_times <- data.frame(
  distance = c(5, 10, 20, 30, 35),
  time = c(1.20, 1.96, 3.36, 4.71, 5.35)
)

# Simple model
simple_model <- with(
  split_times,
  model_using_splits(distance, time)
)

head(predict(simple_model))
}
