% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts_S3.R
\name{print.shorts_mixed_model}
\alias{print.shorts_mixed_model}
\title{S3 method for printing \code{shorts_mixed_model} object}
\usage{
\method{print}{shorts_mixed_model}(x, ...)
}
\arguments{
\item{x}{\code{shorts_mixed_model} object}

\item{...}{Not used}
}
\description{
S3 method for printing \code{shorts_mixed_model} object
}
\examples{
data("split_times")

mixed_model <- mixed_model_using_splits(
  data = split_times,
  distance = "distance",
  time = "time",
  athlete = "athlete"
)

print(mixed_model)
}
