% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts_S3.R
\name{print.shorts_model}
\alias{print.shorts_model}
\title{S3 method for printing \code{shorts_model} object}
\usage{
\method{print}{shorts_model}(x, ...)
}
\arguments{
\item{x}{\code{shorts_model} object}

\item{...}{Not used}
}
\description{
S3 method for printing \code{shorts_model} object
}
\examples{
split_times <- data.frame(
  distance = c(5, 10, 20, 30, 35),
  time = c(1.20, 1.96, 3.36, 4.71, 5.35)
)

# Simple model
simple_model <- with(
  split_times,
  model_using_splits(distance, time)
)

print(simple_model)
}
