% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-radar-gun.R
\docType{data}
\name{radar_gun_data}
\alias{radar_gun_data}
\title{Radar Gun Data}
\format{
Data frame with 4 variables and 3000 observations:
\describe{
   \item{athlete}{Character string}
   \item{bodyweight}{Bodyweight in kilograms}
   \item{time}{Time reported by the radar gun in seconds}
   \item{velocity}{Velocity reported by the radar gun in m/s}
}
}
\usage{
data(radar_gun_data)
}
\description{
Data generated from known MSS and TAU and measurement error for N=5 athletes using radar gun
    with sampling frequency of 100Hz over 6 seconds.
}
\keyword{datasets}
