% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts-S3.R
\name{residuals.shorts_model}
\alias{residuals.shorts_model}
\title{S3 method for returning residuals of \code{shorts_model}}
\usage{
\method{residuals}{shorts_model}(object, ...)
}
\arguments{
\item{object}{\code{shorts_model} object}

\item{...}{Extra arguments. Not used}
}
\description{
S3 method for returning residuals of \code{shorts_model}
}
\examples{
split_distances <- c(10, 20, 30, 40, 50)
split_times <- create_timing_gates_splits(
  gates = split_distances,
  MSS = 10,
  MAC = 9,
  FD = 0.25,
  TC = 0
)

# Simple model
simple_model <- model_timing_gates(split_distances, split_times)
residuals(simple_model)
}
