% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_times.R
\docType{data}
\name{split_times}
\alias{split_times}
\title{Split Testing Data}
\format{An object of class \code{data.frame} with 30 rows and 4 columns.}
\usage{
data(split_times)
}
\description{
Data generated from known MSS and TAU and measurement error for 5 athletes using 6 timing gates:
    5m, 10m, 15m, 20m, 30m, 40m
}
\details{
\describe{
   \item{athlete}{Character string}
   \item{bodyweight}{Bodyweight in kilograms}
   \item{distance}{Distance of the timing gates from the sprint start in meters}
   \item{time}{Time reported by the timing gate}
}
}
\keyword{datasets}
