% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts-S3.R
\name{coef.shorts_model}
\alias{coef.shorts_model}
\title{S3 method for extracting model parameters from \code{shorts_model} object}
\usage{
\method{coef}{shorts_model}(object, ...)
}
\arguments{
\item{object}{\code{shorts_model} object}

\item{...}{Extra arguments. Not used}
}
\description{
S3 method for extracting model parameters from \code{shorts_model} object
}
\examples{
split_distances <- c(10, 20, 30, 40, 50)
split_times <- create_timing_gates_splits(
  gates = split_distances,
  MSS = 10,
  MAC = 9,
  FD = 0.25,
  TC = 0
)

# Simple model
simple_model <- model_timing_gates(split_distances, split_times)
coef(simple_model)
}
