% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts_S3.R
\name{print.shorts_fv_profile}
\alias{print.shorts_fv_profile}
\title{S3 method for printing \code{shorts_fv_profile} object}
\usage{
\method{print}{shorts_fv_profile}(x, ...)
}
\arguments{
\item{x}{\code{shorts_fv_profile} object}

\item{...}{Not used}
}
\description{
S3 method for printing \code{shorts_fv_profile} object
}
\examples{
data("jb_morin")

m1 <- model_using_radar_with_time_correction(time = jb_morin$time, velocity = jb_morin$velocity)

fv_profile <- get_FV_profile(
  MSS = m1$parameters$MSS,
  TAU = m1$parameters$TAU,
  bodyheight = 1.72,
  bodymass = 120)

print(fv_profile)
}
