% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-split-times.R
\docType{data}
\name{split_times}
\alias{split_times}
\title{Split Testing Data}
\format{
Data frame with 4 variables and 30 observations:
\describe{
   \item{athlete}{Character string}
   \item{bodyweight}{Bodyweight in kilograms}
   \item{distance}{Distance of the timing gates from the sprint start in meters}
   \item{time}{Time reported by the timing gate}
}
}
\usage{
data(split_times)
}
\description{
Data generated from known MSS and TAU and measurement error for N=5 athletes using 6 timing gates:
    5m, 10m, 15m, 20m, 30m, 40m
}
\keyword{datasets}
